clean code structure
