+++ src/components/CarouselCardSkeleton.tsx
@@ -1,12 +1,16 @@
-import "../css/ProjectCarouselSkeleton.css"
+import "../css/CarouselCardSkeleton.css"
 
-const ProjectCarouselCardSkeleton: React.FC = () => {
+const CarouselCardSkeleton: React.FC = () => {
 
     return (
-        <div className="ProjectCarouselCard">
+        <div className="SkeletonCard">
             <div className="card-top">
                 <div className="git-icon skeleton"></div>
-                <p className="project-owner skeleton skeleton-text">&nbsp;</p>
+                <div className="owner-name">
+                    <div className="skeleton skeleton-text"></div>
+                    <div className="skeleton skeleton-text"></div>
+                </div>
+                
             </div>
             <div className="project-owner-image skeleton"></div>
             <div className="card-bottom">
@@ -15,27 +19,13 @@ const ProjectCarouselCardSkeleton: React.FC = () => {
                         <svg className="skeleton-stroke" height="16" width="16" viewBox="0 0 16 16">
                             <path d="M2 2.5A2.5 2.5 0 0 1 4.5 0h8.75a.75.75 0 0 1 .75.75v12.5a.75.75 0 0 1-.75.75h-2.5a.75.75 0 0 1 0-1.5h1.75v-2h-8a1 1 0 0 0-.714 1.7.75.75 0 1 1-1.072 1.05A2.495 2.495 0 0 1 2 11.5Zm10.5-1h-8a1 1 0 0 0-1 1v6.708A2.486 2.486 0 0 1 4.5 9h8ZM5 12.25a.25.25 0 0 1 .25-.25h3.5a.25.25 0 0 1 .25.25v3.25a.25.25 0 0 1-.4.2l-1.45-1.087a.249.249 0 0 0-.3 0L5.4 15.7a.25.25 0 0 1-.4-.2Z"></path>
                         </svg>
-                        <p className="skeleton skeleton-text">
-                            &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;
-                            &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;
-                            &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;
-                        </p>
+                        <p className="skeleton skeleton-text"></p>
                     </div>
 
                         <div className="about-project">
-                            <div className="skeleton skeleton-text">
-                                &nbsp;&nbsp;&nbsp;&nbsp;
-                                &nbsp;&nbsp;&nbsp;&nbsp;
-                                &nbsp;&nbsp;&nbsp;&nbsp;
-                            </div>
-                            <p className="skeleton skeleton-text">
-                                &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;
-                                &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;
-                                &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;
-                                &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;
-                                &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;
-                                &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;
-                            </p>
+                            <p className="skeleton skeleton-text"></p>
+                            <p className="skeleton skeleton-text"></p>
+                            <p className="skeleton skeleton-text"></p>
                         </div>
 
                 </div>
@@ -67,4 +57,4 @@ const ProjectCarouselCardSkeleton: React.FC = () => {
     )
 }
 
-export default ProjectCarouselCardSkeleton
+export default CarouselCardSkeleton
