+++ src/css/CarouselCardComplete.css
@@ -1,4 +1,4 @@
-.ProjectCarouselCard {
+.CarouselCardComplete {
     width: var(--carousel-card-width);
     height: 400px;
     border: 1px solid #333;
@@ -9,7 +9,7 @@
     flex-shrink: 0;
 }
 
-.ProjectCarouselCard:hover {
+.CarouselCardComplete:hover {
     transform: scale(1.1);
     margin: 0px 20px 0px 20px;
 }
@@ -132,36 +132,4 @@
     width: 12px;
     height: 12px;
     border-radius: 50%;
-}
-
-.skeleton {
-    animation: skeleton-loading 1s linear infinite alternate;
-}
-
-.skeleton-stroke {
-    animation: skeleton-stroke-loading 1s linear infinite alternate;
-}
-
-.skeleton-text {
-    border-radius: .125rem;
-}
-  
-@keyframes skeleton-loading {
-    0% {
-      background-color: hsl(198, 10%, 53%);
-    }
-  
-    100% {
-      background-color: hsl(0, 0%, 35%);
-    }
-}
-
-@keyframes skeleton-stroke-loading {
-    0% {
-      stroke: hsl(198, 10%, 53%);
-    }
-  
-    100% {
-        stroke: hsl(0, 0%, 35%);
-    }
 }
