+++ src/css/CarouselCardSkeleton.css
@@ -1,4 +1,4 @@
-.ProjectCarouselCard {
+.SkeletonCard {
     width: var(--carousel-card-width);
     height: 400px;
     border: 1px solid #333;
@@ -9,12 +9,7 @@
     flex-shrink: 0;
 }
 
-.ProjectCarouselCard:hover {
-    transform: scale(1.1);
-    margin: 0px 20px 0px 20px;
-}
-
-.card-top {
+.SkeletonCard .card-top {
     height: 30%;
     width: 100%;
     display: flex;
@@ -25,14 +20,14 @@
     border-bottom: 2px solid black;
 }
 
-.git-icon {
+.SkeletonCard .git-icon {
     width: 30px;
     height: 30px;
     border-radius: 50%;
     margin-left: auto;
 }
 
-.project-owner-image {
+.SkeletonCard .project-owner-image {
     width: 80px;
     height: 80px;
     border: 1px solid black;
@@ -43,20 +38,15 @@
     transform: translate(0px, -70%);
 }
 
-.project-owner {
-    text-align: left;
-    margin: 0px 0px 0px calc(20px + 80px);
-    overflow: hidden;
-    text-overflow: ellipsis;
-    word-wrap: break-word;
-    display: -webkit-box;
-    -webkit-line-clamp: 2;
-    line-clamp: 2;
-    -webkit-box-orient: vertical;
-    color: white;
+.SkeletonCard .owner-name {
+    display: flex;
+    flex-direction: column;
+    margin: 0px 0px 0px 100px;
+    width: calc(100%-100px);
+    gap: 2px;
 }
 
-.card-bottom {
+.SkeletonCard .card-bottom {
     height: 70%;
     width: 100%;
     display: flex;
@@ -68,27 +58,14 @@
     text-align: left;
 }
 
-.project-name {
+.SkeletonCard .project-name {
     display: flex;
     align-items: center;
     gap: 5px;
     margin-bottom: 10px;
 }
 
-.project-name path {
-    fill: white;
-}
-
-.project-name a {
-    text-decoration: none;
-    color: #4493f8;
-}
-
-.project-name a:hover {
-    text-decoration: underline;
-}
-
-.about-project {
+.SkeletonCard .about-project {
     display: flex;
     flex-direction: column;
     align-items: start;
@@ -96,18 +73,7 @@
     font-size: 14px;
 }
 
-.about-project p {
-    margin: 5px 0px 5px 0px;
-    overflow: hidden;
-    text-overflow: ellipsis;
-    word-wrap: break-word;
-    display: -webkit-box;
-    -webkit-line-clamp: 3;
-    line-clamp: 2;
-    -webkit-box-orient: vertical;
-}
-
-.project-stats {
+.SkeletonCard .project-stats {
     width: 100%;
     display: flex;
     justify-content: start;
@@ -116,20 +82,57 @@
     font-size: 12px;
 }
 
-.project-stat {
+.SkeletonCard .project-stat {
     display: flex;
     flex-direction: row;
     align-items: center;
     gap: 2px;
 }
 
-.project-stats path {
+.SkeletonCard .project-stats path {
     fill: #9198a1;
 }
 
 .repo-language-color {
-    background-color: #3572A5;
     width: 12px;
     height: 12px;
     border-radius: 50%;
+}
+
+.skeleton {
+    animation: skeleton-loading 1s linear infinite alternate;
+}
+
+.skeleton-stroke {
+    animation: skeleton-stroke-loading 1s linear infinite alternate;
+}
+
+.skeleton-text {
+    width: 100%;
+    height: 12px;
+    border-radius: .125rem;
+}
+
+.skeleton-text:last-child {
+    width: 70%;
+}
+  
+@keyframes skeleton-loading {
+    0% {
+      background-color: hsl(198, 10%, 53%);
+    }
+  
+    100% {
+      background-color: hsl(0, 0%, 35%);
+    }
+}
+
+@keyframes skeleton-stroke-loading {
+    0% {
+      stroke: hsl(198, 10%, 53%);
+    }
+  
+    100% {
+        stroke: hsl(0, 0%, 35%);
+    }
 }
